Skeleton of external credit rewards
