+++ kifi-backend/modules/shoebox/app/com/keepit/payments/DescriptionElements.scala
@@ -1,0 +1,70 @@
+package com.keepit.payments
+
+import com.keepit.common.mail.EmailAddress
+import com.keepit.common.path.Path
+import com.keepit.model.{ OrganizationRole, OrganizationHandle, BasicOrganization }
+import com.keepit.social.BasicUser
+import play.api.libs.json.{ Writes, Json }
+import play.twirl.api.Html
+
+sealed trait DescriptionElements {
+  def flatten: Seq[BasicElement]
+}
+case class SequenceOfElements(elements: Seq[DescriptionElements]) extends DescriptionElements {
+  def flatten = elements.flatMap(_.flatten)
+}
+case class BasicElement(text: String, url: Option[String]) extends DescriptionElements {
+  def flatten = Seq(this)
+  def withText(newText: String) = this.copy(text = newText)
+}
+object DescriptionElements {
+  def apply(elements: DescriptionElements*): SequenceOfElements = SequenceOfElements(elements)
+
+  implicit def fromText(text: String): BasicElement = BasicElement(text, None)
+  implicit def fromTextAndUrl(textAndUrl: (String, String)): BasicElement = BasicElement(textAndUrl._1, Some(textAndUrl._2))
+  implicit def fromSeq[T](seq: Seq[T])(implicit toElements: T => DescriptionElements): SequenceOfElements = SequenceOfElements(seq.map(toElements))
+  implicit def fromOption[T](opt: Option[T])(implicit toElements: T => DescriptionElements): SequenceOfElements = opt.toSeq
+
+  implicit def fromCreditCode(code: CreditCode): BasicElement = code.value
+  implicit def fromBasicUser(user: BasicUser): BasicElement = user.firstName -> user.path.absolute
+  implicit def fromBasicOrg(org: BasicOrganization): BasicElement = org.name -> org.path.absolute
+  implicit def fromEmailAddress(email: EmailAddress): BasicElement = email.address
+  implicit def fromDollarAmount(v: DollarAmount): BasicElement = v.toDollarString
+  implicit def fromPaidPlanAndUrl(plan: PaidPlan)(implicit orgHandle: OrganizationHandle): BasicElement = plan.fullName -> Path(s"${orgHandle.value}/settings/plan").absolute
+  implicit def fromRole(role: OrganizationRole): BasicElement = role.value
+
+  private def intersperse[T](xs: List[T], ins: List[T]): List[T] = {
+    (xs, ins) match {
+      case (x :: Nil, Nil) => x :: Nil
+      case (x :: xr, in :: inr) => x :: in :: intersperse(xr, inr)
+      case _ => throw new IllegalArgumentException(s"intersperse expects lists with length (n, n-1). it got (${xs.length}, ${ins.length})")
+    }
+  }
+  private def interpolatePunctuation(els: Seq[BasicElement]): Seq[BasicElement] = {
+    val words = els.map(_.text).toList
+    val wordPairs = words.init zip words.tail
+    val interpolatedPunctuation = wordPairs.map {
+      case (l, r) if l.endsWith("'") || r.startsWith(".") || r.startsWith("'") => ""
+      case _ => " "
+    }.map(BasicElement(_, None))
+    intersperse(els.toList, interpolatedPunctuation).filter(_.text.nonEmpty)
+  }
+  def formatForSlack(description: DescriptionElements): String = {
+    interpolatePunctuation(description.flatten).map {
+      case BasicElement(text, None) => text
+      case BasicElement(text, Some(url)) => s"<$url|$text>"
+    } mkString
+  }
+
+  def formatAsHtml(description: DescriptionElements): Html = {
+    Html(interpolatePunctuation(description.flatten).map {
+      case BasicElement(text, None) => text
+      case BasicElement(text, Some(url)) => s"""<a href="$url">$text</a>"""
+    } mkString)
+  }
+
+  implicit val flatWrites = {
+    implicit val basicWrites = Json.writes[BasicElement]
+    Writes[DescriptionElements] { description => Json.toJson(interpolatePunctuation(description.flatten)) }
+  }
+}
